+++ src/permission-observer.ts
@@ -1,0 +1,49 @@
+export type PermissionObserver = {
+  waitForState: (
+    states: PermissionState | NonEmptyPermissionStateArray,
+    task?: () => Promise<void>,
+  ) => Promise<void>;
+  waitForStateChange: (
+    states?: PermissionState | PermissionState[],
+    task?: () => Promise<void>,
+  ) => Promise<void>;
+};
+
+export async function createPermissionObserver(
+  permissions: Permissions,
+  descriptor: PermissionDescriptor,
+): Promise<PermissionObserver> {
+  const status = await permissions.query(descriptor);
+
+  return {
+    async waitForState(states, task) {
+      if (states.length < 1) throw new Error("No states provided");
+      if (states.includes(status.state)) return;
+
+      await waitForStateChange(states, task);
+    },
+
+    waitForStateChange,
+  };
+
+  async function waitForStateChange(
+    states: PermissionState | PermissionState[] = [],
+    task?: () => Promise<void>,
+  ) {
+    await Promise.all([
+      new Promise<void>((resolve) => {
+        status.addEventListener("change", onChange);
+
+        function onChange() {
+          if (states.length > 0 && !states.includes(status.state)) return;
+
+          status.removeEventListener("change", onChange);
+          resolve();
+        }
+      }),
+      Promise.resolve(task?.()),
+    ]);
+  }
+}
+
+type NonEmptyPermissionStateArray = PermissionState[] & { 0: PermissionState };